+++ src/main/java/com/oya/kr/popup/controller/dto/response/PopupListResponse.java
@@ -1,5 +1,6 @@
 package com.oya.kr.popup.controller.dto.response;
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 import com.oya.kr.popup.domain.Category;
@@ -16,7 +17,8 @@ public class PopupListResponse {
     private final String title;
     private final String description;
     private final LocalDateTime pulledDate;
-    private final LocalDateTime createdDate;
+    private final LocalDate openDate;
+    private final LocalDate closeDate;
     private final String thumbnail;
     private final CategoryResponse category;
 
@@ -26,7 +28,8 @@ public class PopupListResponse {
             response.getTitle(),
             response.getDescription(),
             response.getPulledDate(),
-            response.getCreatedDate(),
+            response.getOpenDate(),
+            response.getCloseDate(),
             response.getThumbnail(),
             CategoryResponse.from(Category.from(response.getCategory()))
         );
@@ -43,14 +46,19 @@ public class PopupListResponse {
             pulledDate.getSecond());
     }
 
-    public String getCreatedDate() {
+    public String getOpenDate() {
         return String.format(
-            "%d-%02d-%dT%02d:%02d:%02d",
-            createdDate.getYear(),
-            createdDate.getMonthValue(),
-            createdDate.getDayOfMonth(),
-            createdDate.getHour(),
-            createdDate.getMinute(),
-            createdDate.getSecond());
+            "%d-%02d-%d",
+            openDate.getYear(),
+            openDate.getMonthValue(),
+            openDate.getDayOfMonth());
+    }
+
+    public String getCloseDate() {
+        return String.format(
+            "%d-%02d-%d",
+            closeDate.getYear(),
+            closeDate.getMonthValue(),
+            closeDate.getDayOfMonth());
     }
 }
